pyproject.toml
[tool.poetry]
name = "fast_zero"
version = "0.1.0"
description = ""
authors = ["RaissaAndradeS <221035077@aluno.unb.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = {extras = ["standard"], version = "^0.115.8"}
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.18"
pyjwt = "^2.10.1"
psycopg = {extras = ["binary"], version = "^3.2.4"}
pandas = "^2.2.0"
numpy = "^1.26.0"
pdfplumber = "0.10.4"
python-magic-bin = "^0.4.14"
PyPDF2 = "^3.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.6"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
factory-boy = "^3.3.3"
freezegun = "^1.5.1"
testcontainers = "^4.9.1"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.taskipy.tasks]
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest --cov=fast_zero -VV'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
